FareStructureElement refactor complete
