+++ repos/fdbt-netex-output/src/netex-convertor/period-tickets/periodTicketNetexHelpers.test.ts
@@ -423,119 +423,6 @@ describe('periodTicketNetexHelpers', () => {
         });
     });
 
-    describe('getFareStructureElements', () => {
-        it('returns 3 fareSructureElements for each product in the products array for multiService; Access Zones, Eligibility and Conditions of Travel', () => {
-            const expectedLength = flatFareTicket.products.length * 3;
-            const result = netexHelpers.getFareStructuresElements(flatFareTicket as PeriodTicket, placeHolderText);
-            const namesOfTypesOfFareStructureElements: string[] = result.map(element => {
-                return element.Name.$t;
-            });
-
-            namesOfTypesOfFareStructureElements.forEach(name => {
-                expect(
-                    name === 'Available lines and/or zones' ||
-                        name === 'Eligible user types' ||
-                        name === 'Conditions of travel',
-                ).toBeTruthy();
-            });
-
-            expect(result.length).toBe(expectedLength);
-        });
-
-        it('returns 3 fareSructureElements for each product in the products array for multiService; Access Zones, Durations and Conditions of Travel and 1 for eligibility', () => {
-            const expectedLength = periodMultipleServicesTicket.products.length * 3 + 1;
-            const result = netexHelpers.getFareStructuresElements(periodMultipleServicesTicket, placeHolderText);
-            const namesOfTypesOfFareStructureElements: string[] = result.map(element => {
-                return element.Name.$t;
-            });
-
-            namesOfTypesOfFareStructureElements.forEach(name => {
-                expect(
-                    name === 'Available lines and/or zones' ||
-                        name.includes('Available duration combination') ||
-                        name === 'Eligible user types' ||
-                        name === 'Conditions of travel',
-                ).toBeTruthy();
-            });
-            expect(result.length).toBe(expectedLength);
-        });
-
-        it('returns 3 fareStructureElements for each product in the products array for geoZone; Access Zones, Durations and Conditions of Travel and 1 for eligibility', () => {
-            const expectedLength = geoUserPeriodTicket.products.length * 3 + 1;
-            const result = netexHelpers.getFareStructuresElements(geoUserPeriodTicket, placeHolderText);
-            const namesOfTypesOfFareStructureElements: string[] = result.map(element => {
-                return element.Name.$t;
-            });
-
-            namesOfTypesOfFareStructureElements.forEach(name => {
-                expect(
-                    name === 'Available lines and/or zones' ||
-                        name.includes('Available duration combination') ||
-                        name === 'Eligible user types' ||
-                        name === 'Conditions of travel',
-                ).toBeTruthy();
-            });
-            expect(result.length).toBe(expectedLength);
-        });
-
-        it('returns the fareStructureElements in the format we expect', () => {
-            const geoResult = netexHelpers.getFareStructuresElements(geoUserPeriodTicket, placeHolderText);
-
-            const expectedAccessZonesFareStructureElement = {
-                version: '1.0',
-                id: expect.stringContaining('op:Tariff@'),
-                Name: expect.objectContaining({ $t: expect.any(String) }),
-                TypeOfFareStructureElementRef: expect.objectContaining({
-                    version: expect.any(String),
-                    ref: expect.any(String),
-                }),
-                GenericParameterAssignment: expect.any(Object),
-                qualityStructureFactors: null,
-            };
-
-            const expectedDurationsFareStructureElement = {
-                version: '1.0',
-                id: expect.stringContaining('op:Tariff@'),
-                Name: expect.objectContaining({ $t: expect.any(String) }),
-                TypeOfFareStructureElementRef: expect.objectContaining({
-                    version: expect.any(String),
-                    ref: expect.any(String),
-                }),
-                timeIntervals: expect.objectContaining({ TimeIntervalRef: expect.anything() }),
-            };
-
-            const expectedEligibilityFareStructureElement = {
-                version: '1.0',
-                id: expect.stringContaining('op:Tariff@'),
-                Name: expect.objectContaining({ $t: expect.any(String) }),
-                TypeOfFareStructureElementRef: expect.objectContaining({
-                    version: expect.any(String),
-                    ref: expect.any(String),
-                }),
-                GenericParameterAssignment: expect.any(Object),
-            };
-
-            const expectedConditionsOfTravelFareStructureElement = {
-                version: '1.0',
-                id: expect.stringContaining('op:Tariff@'),
-                Name: expect.objectContaining({ $t: expect.any(String) }),
-                GenericParameterAssignment: expect.any(Object),
-            };
-
-            geoResult.forEach(fareStructureElement => {
-                if (fareStructureElement.timeIntervals) {
-                    expect(fareStructureElement).toEqual(expectedDurationsFareStructureElement);
-                } else if (fareStructureElement.qualityStructureFactors === null) {
-                    expect(fareStructureElement).toEqual(expectedAccessZonesFareStructureElement);
-                } else if (fareStructureElement.TypeOfFareStructureElementRef) {
-                    expect(fareStructureElement).toEqual(expectedEligibilityFareStructureElement);
-                } else {
-                    expect(fareStructureElement).toEqual(expectedConditionsOfTravelFareStructureElement);
-                }
-            });
-        });
-    });
-
     describe('getGroupOfOperators', () => {
         beforeEach(() => {
             jest.spyOn(db, 'getOperatorDataByNocCode').mockImplementationOnce(() => Promise.resolve(multiOperatorList));
